feat(dashboardUsers) : added getUserDetails, unlockUser, update user role endpoints with their service methods , added new role 'SuperAdmin' and modified all previous endpoint accordingly, added new permissions service to use to check if the current logged in user has permissions to act on another target user, modified users seeding accordingly.

+++ ECommerce.API/Controllers/UsersController.cs
@@ -14,9 +14,10 @@ namespace ECommerce.API.Controllers
     [ApiController]
     [Tags("Users Management")]
     [Authorize]
-    public class UsersController(IUserService users) : ControllerBase
+    public class UsersController(IUserService users, IPermissionService permissions) : ControllerBase
     {
         private readonly IUserService _users = users;
+        private readonly IPermissionService _permissions = permissions;
 
 
         [HttpGet]
@@ -82,7 +83,7 @@ namespace ECommerce.API.Controllers
 
         }
 
-        [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "SuperAdmin,Admin")]
         [HttpGet("admin")]
         [EndpointSummary("Admin lists all users with filter,search,sort and pagination support.")]
         [ProducesResponseType(typeof(PagedResponse<AdminUserSummaryDto>), StatusCodes.Status200OK)]
@@ -96,7 +97,40 @@ namespace ECommerce.API.Controllers
             return Ok(response);
         }
 
-        [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "SuperAdmin,Admin")] //SuperAdmin -> Admin,Seller,Customer , Admin -> Seller, Customer
+        [HttpGet("admin/{userId:guid}")]
+        [EndpointSummary("Admin retrieves a certain user details.")]
+        [ProducesResponseType(typeof(AdminUserDetailsResponse), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status404NotFound)]
+        public async Task<IActionResult> GetUserDetailsAdmin([FromRoute] string userId)
+        {
+            if (!await _permissions.CanManageUserAsync(userId))
+                return Forbid();
+            var user = await _users.GetUserDetailsAdminAsync(userId);
+            return Ok(user);
+        }
+
+        [Authorize(Roles = "SuperAdmin,Admin")] //SuperAdmin -> Admin,Seller,Customer , Admin -> Seller, Customer
+        [HttpPut("admin/{userId:guid}")]
+        [EndpointSummary("Admin updates user's role.")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status404NotFound)]
+        public async Task<IActionResult> UpdateUserRoleAdmin([FromRoute] string userId, [FromBody] AdminUpdateRoleRequest adminUpdateRoleRequest)
+        {
+            if (!await _permissions.CanManageUserAsync(userId))
+                return Forbid();
+            await _users.UpdateUserRoleAdminAsync(userId, adminUpdateRoleRequest);
+            return Ok();
+        }
+
+
+        [Authorize(Roles = "SuperAdmin,Admin")] //SuperAdmin -> Admin,Seller,Customer , Admin -> Seller,Customer
         [HttpDelete("admin/{userId:guid}")]
         [EndpointSummary("Admin soft deletes a user account.")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
@@ -106,11 +140,13 @@ namespace ECommerce.API.Controllers
         [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status404NotFound)]
         public async Task<IActionResult> DeleteUserAdmin([FromRoute] string userId)
         {
+            if (!await _permissions.CanManageUserAsync(userId))
+                return Forbid();
             await _users.DeleteUserAdminAsync(userId);
             return NoContent();
         }
 
-        [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "SuperAdmin,Admin")] //SuperAdmin -> Admin,Seller,Customer , Admin -> Seller,Customer
         [HttpPut("admin/{userId:guid}/restore")]
         [EndpointSummary("Admin restores a deleted user account.")]
         [ProducesResponseType(StatusCodes.Status200OK)]
@@ -120,7 +156,25 @@ namespace ECommerce.API.Controllers
         [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status404NotFound)]
         public async Task<IActionResult> RestoreDeletedUserAdmin([FromRoute] string userId)
         {
+            if (!await _permissions.CanManageUserAsync(userId))
+                return Forbid();
             await _users.RestoreDeletedUserAdminAsync(userId);
+            return Ok();
+        }
+
+        [Authorize(Roles = "SuperAdmin,Admin")] //SuperAdmin -> Admin,Seller,Customer , Admin -> Seller,Customer
+        [HttpPut("admin/{userId:guid}/unlock")]
+        [EndpointSummary("Admin Instantly unlock a locked user account.")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status404NotFound)]
+        public async Task<IActionResult> UnlockUserAdmin([FromRoute] string userId)
+        {
+            if (!await _permissions.CanManageUserAsync(userId))
+                return Forbid();
+            await _users.UnlockUserAdminAsync(userId);
             return Ok();
         }
